Remove unused code and rename functions that were called "Local"
In the past, functions were named with the word "Local" in them if it was
anticipated that we would need to write a separate function for global
deliveries. This is no longer true for some functions, which have been renamed.

getLocalTargets has been replaced by getParserFunctionTargets and is no longer
used at all, and can safely be removed.

Change-Id: I23f22a317060bd69110b2a3dda2de396bad65cf5