Throw errors if user input for the parser function is not valid
If the user provides a site, check that it is a valid url. For the page
title the user provides, ensure that it is a valid page title.

An error message is returned rather than the normal output, and that
page/site combo will not be added to the list of targets.

Change-Id: I115955de5425becbb4798bb0a2c15ed6e7f37ff3