Add debug logging for bug 57464
Currently any namespace that is equal to an interwiki
prefix on the submission site is mysteriously dropped.

1dd6eab3578 should have fixed this, but it didn't.
If the debug statements show that the interwiki prefix
is still retained after a new Title object is created, 
we will know that the previous fix was inadequate.
If not, something else is terribly wrong.

Bug: 57464
Change-Id: Iec240623ae73b647ed700fb5ce0ff3277152ddc7
(cherry picked from commit 8d28ff418e7b20fa129528072041ec655c5cf517)