Streamline ListContentSpamlistLookupTest
This demonstrates some possible methods to reduce the complexity of the
test code introduced in Ie766c43.

The class under test (ListContentSpamlistLookup) is pretty simple, and so
is it's abstract base class (SpamlistLookup). Listing the covered methods
individually is more confusing than helpful. Accidential coverage is not
that much of a problem for classes that do one very specific thing, as
these do.

Change-Id: Id45c886f5a909191e16145edce7cd61cff3e20d2