Avoid hard coding namespaces, use the ::class constant instead
__CLASS__ can also be used in this case but ::class seems better
and shorter. Avoiding to hard code the namespaces here is good as
even if the namespaces change, ::class should be able to resolve
to the correct namespaces and this improve code readability.

Except the class names change that is when we can touch something
here but it's very unlikely that the class names will change.

Change-Id: I78b4b6620898e41c360dc25a2a0ba2386ff62023