MassMessageJob: Set '$this->title' after constructor call
The previous code was technically ambiguous in that it worked
really hard to create a valid $title parameter, but effectively
didn't want it to be used. It wants $params['title' to survive.

This conflicts with fc5d51f12 from mediawiki/core, which started
honouring the passed in $title (or tried to), but failed to do so
properly due to $params['title'] also existing, thus leading to
the wrong conclusion that the logic had already run and we're now
in execution mode instead of queueing mode. (thus $title is not
stored, but $params['title'] - while stored, also not used to create
$this->title because $params['namespace'] is missing).

Fix this by effectively formalising that this job wants $params
to be stored as-is, and that $this->title is purely a run-time
concept. For later, this should perhaps use a different instance
member name so that this is clearly a locally owned member that
must be set and not relying on parent class logic which is for
something different technically.

Bug: T221570
Change-Id: I18026b19e586b81f60bc95fc18705905d491ab39