Allow editing spamlist descriptions using the API
When editing a spamlist using the API, a `description` parameter
can be passed in addition to or in lieu of adding or removing
entries. If the description is identical to the current spamlist
description, the result of the API call will be `Done` instead
of `Success`, with the description returned as `invaliddescription`.
If the description is saved, it will be returned as `description`.
The API is a post module, and requires the standard csrf token.

Bug: T226929
Change-Id: I09ac96a82e98fa6da40738a54dbf4b0da1bd8827