Use new PermissionManager::addTemporaryUserRights instead of hook
The previous hook was only added at runtime and fragile, since rights
might have already been cached by that point. Now that there's a proper
helper for it, use that.

Test plan was to hardblock 127.0.0.1 and ::1, then send a MassMessage,
and see that they are still delivered properly.

And silence new PHPCS warnings temporarily due to bumping MediaWiki
version requirement.

Bug: T228252
Change-Id: Ic4f8f64b523b6823ae90fdd57108067a362e1474