Provide deduplication info to MassMessageJob
The MassMessageSubmitJob is pushing jobs in batches by wiki.
If the job fails for one of the wikis, it gets retried, which
creates duplicate messages for wikis where the submit succeeded.

Setting removeDuplicates flag to true will allow the deduplication
of the messages while they are in the queue, however, most likely
the previous duplicated job was already executed when a submit job
is retried, thus removeDuplicates flag is not enough. Deduplicating
but the root job will work - it removes duplicate jobs that originate
from the same root job if both duplicates were posted after the original
root job was enqueued - that's exactly what we need.

Bug: T232379
Change-Id: Idc8493c0cae201ca8e0fe2f0c38a0baaa38a5660
(cherry picked from commit bea9564e37aa6efa0813bd85b739c7ee2534539e)