Properly return API failure to MassMessageJob
In case of a failure we do several attempts to call the API,
but if all of them end in an ApiUsageException we give up.

However, in the internal mode, which is used here, checking
for an error ley in the result is not enough, since it will
not be set.

Bug: T246217
Change-Id: I8a22e231cc4669ff8b59276217d24d9564c78b3f