Add a page selector on Special:MassMessage to be sent as message
This can be the title of any page in the wiki. If provided this will
be appended to the message entered by the sender and sent to
subscribers. It is mandatory to either provide this or the message.

If the selected page is a source translation page, an info message
will be shown to the sender upon preview stating that subscribers will
receive the message in the translated form as per their preference.

This patch deals with changes for the UI and API to accept the
page to be sent as message, and displaying that page on preview.

Bug: T165128
Change-Id: Idf2ffb59b4cac4ba3d32b9f03c83434e9ffe03f8