Add support for sending only the specified section of a page
This works for both non-translatable and translatable pages.
I did look at LST extension, but it cannot be reused here without
a lot of refactoring.

This re-implements LST functionality by parsing section tags
directly using regular expressions. It does not currently
support i18ned tags, but that is not an issue for TechNews.

Bug: T254483
Change-Id: If8cfe16e7c875f001b1477323e4eeba52c59fff1