Remove generate-html false optimization
Saving a couple ms here on a non-hot path doesn't seem worth it.
Additionally, core requires that if you don't generate html that
the ParserOutput->mText field be null, however ParserOutput::setText
refuses to set it to null, which makes it difficult to comply
with the requirements

Not generating html is an optional optimization, so it is always
safe to just fully generate html.

Bug: T299896
Change-Id: I7c32b77e0541e03dafa6afc6eeccacc5de86eb81