Convert Special:MassMessage to use OOUI
Convert Special:MassMessage to use OOUI. This includes making the
Special page extend FormSpecialPage to reduce the amount of code
and also modifying the JS validation to fit with the OOUI style.
CSS and JS is also updated to reflect changing to OOUI which causes
classes and IDs to change.

The title fields were also updated to the title field type to match
with the OOUI styling as the autocomplete JS code currently used
does not match OOUI styling. One downside of this is that the
suggested pages can include pages that are not valid options.
A change to the OOUI title field is likely needed for this.

Also remove uses of $.debounce and replace it with OO.ui.debounce
as one of the two uses needed to be modified to update to OOUI.

Bug: T104419
Bug: T213426
Change-Id: Ie37fdbe04b4ae85d3f3020d943f41fbe1c477094