content: Port frontend to OOUI, drop jquery.ui
This removes the final usage of jquery.ui in this extension.

The form is now built using OOUI PHP, then infused client-side. We use a
TitleInputWidget for the title, and a ComboBoxInputWidget for the site.
The site input now has all the options preloaded, so we can drop the
list=mmsites API module too.

If there's only local messaging, the title input uses a
ActionFieldLayout to combine the submit button. Otherwise it uses a
plain, stacked FormLayout for the two inputs and button.

On the client side, various DOM operations are now replaced by OOUI
functions. Notably the error handling now sets the error on one of the
two input fields rather than for the form as a whole.

Because the HTML structure has dramatically changed, we use the
RejectParserCacheValue hook to clear any cache entries that don't have
the "ext.MassMessage.content" module (it was previously only added in
the BeforePageDisplay hook)

== Test plan ==
* Type in the title input, get autocomplete results
* Sucessfully add a valid title to the list
* Entering an invalid title ("<") results in an error message
* Entering a title already on the list results in an error message
* Enable global messaging (and hack the return value of
  DatabaseLookup::getDatabases() if necessary), verify site input
  appears in form with autocomplete
* Successfully add a valid title+site to the list
* Entering an invalid site results in an error message
* Disable JavaScript, no forms should appear

For parser cache invalidation, check out the commit before this one and
create a MassMessage list, observe that it uses an old-style form. Look
in the HTML source and look for a comment like:
  <!-- Saved in parser cache with key...

Then checkout this patch and observe that it correctly switches to the
new OOUI form.

Bug: T104419
Bug: T316987
Change-Id: Ic6746b7edee735a94e6b33f9b8048f6af387a482