Add a config that marks targets as pre-validated for message delivery
Add a config, named $wgAllowlistedMassMessageTargets, that allows
bypassing all but the MassMessage exclude category check for
specified pages. Pages are specified using their page ID in this
config, which makes it resistant to page moves.

The intended use is to allow pages to be posted to that are in
namespaces that otherwise are disallowed.

Bug: T325290
Change-Id: I47f31d68ab758c3b548faf37c3a0ed87fdf641c8