Only assign bot to MassMessage user if creating or stealing the account
Only assign the bot user group to the MassMessage user if the code
is stealing or creating the system user. If the user is already a
system user, then don't attempt to assign the bot group so that
wikis can define whether they want the bot flag or not using
Special:UserRights.

Bug: T120832
Change-Id: Ie13e0577e2e308c09a9605e10b1612238ae7bf34