Add a hook to override how a message is sent to a specific page
This particular point in the code already has hard-coded detection
for two specific extensions (LQT and flow). There's also an additional
extension - MassMessageEmail which instructs users to patch this
point of MassMessage. It seems clear that this is a code point that
extensions need to be able to override.

Also bump version since people may want to know if the hook is
present based on the version number.

Change-Id: I2de7cdbba836a0de239f4e72b0063f8a2701e83e