Improve error logentry when subject section is invalid
If that section has a spurious "<" character in the beginning, then an overeager
sanitization with PHP's `strip_tags` will cause the subject to be empty,
which will cause the delivery to Flog threads fail and even normal
discussion pages will miss their section header.

The ideal solution would be to extract the logic to extract the logic to
extract and sanitize the subject and message content into its own class,
and then to reuse that also in the preview step. Then we could already
fail in the preview step and the user-error would be very legible.

However, due to lack of time to implement this solution, this seems like
the next best thing, hopefully making it easier to diagnose.

Bug: T375828
Change-Id: I361714259af66bc042eb49a03bd2d4b96baad61d