Implicitly invalidate the urltodb cache when the wiki list changes
WikimediaMaintenance addWiki.php has a hack for MassMessage which tries
to clear and regenerate the urltodb cache when a wiki is added.

Instead, just hash the list of all wikis and use it in the cache key.
That way, the cache is implicitly cleared when the configuration
changes.

I profiled cache misses. They take about 5ms. 70% of that is from
SiteConfiguration::get(). getLocalDatabases() is precomputed and so is
very fast.

Test log:

* With production config, MassMessage master, removed aawiki from
  all.dblist
* With eval.php print \MediaWiki\MassMessage\Lookup\DatabaseLookup
  ::getDBName('aa.wikipedia.org') -> empty
* Add aawiki to all.dblist
* getDBName() -> still empty. Bug confirmed.
* Remove aawiki, check out MassMessage proposed
* getDBName() -> empty
* Add aawiki
* getDBName() -> aawiki. Fix confirmed.

Bug: T62075
Bug: T352113
Change-Id: I38d334e03a291425357bba2134a06a2a36d96e4b