notify message recipients when a redirect was followed
Appends "(This message was sent to User talk:ABC and is being
posted here due to a redirect.)" to messages that went through
a redirection.

The idea is to not confuse users when a message is intended for
their alt account, but their alt account's user talk redirects to their
main account's user talk page.

Testing notes:
- Messaging to user talk pages of users that don't exist doesn't work.
  Make sure to create the users.
- Posting identical mass messages two in a row doesn't work. It has
  some kind of duplication prevention. Vary your test messages.

Bug: T394413
Change-Id: I8a16a98896743e3714db2bec97894ca29a1bb133